feat(/*): Webpack 4 Support for Universal CSS Chunks
Webpack 4 support for extract-css-chunks universal-component react-universal. Complete with full
chunking, code splitting, real and true hot module reloading. Compatible with faceyspacey universal
code splitting and server-side rendering.

BREAKING CHANGE: EVERYTHING! This its a complete rewrite. Top to bottom!! This is a major change to
the API and Loader. It will intergrate with universal without a issue and  IT DOES NOT RELY ON
UNIVERSAL COMPONENTS.

fix #60, re #36 #47 #51