fix(index.js): Improved HMR detection multi instance collisions addressed
Merging a PR which improved upon my initial code which allowed the plugin to automatically hot
reload. With the help of React Static, we are able to fix a bug that appeared when building out
build toolchains. This bug causes the pugin to have two seperate context references within webpack
and ends up passing incorrect references to the wrong instance. Deceptivally simple solution was
naming the context key myself.