fix: CSS HMR Issues (#114)
* fix: Fixed HMR issue when oneOf is used

When using oneOf, HRM is never activated. This patch improved the lookup system used for determining
if CSS can and should be reloaded. There have been some other tweaks for css modules and HMR

fix #98,#80

* fix: Applying rollback fixes to updateWebpackConfig, search loader and use object keys

* fix: Adding back exported out hot loader

* docs(readme): Adding example of manual hotloader implementation

In the event users cannot get it to work, theres an example in the docs to reference

* fix: handle functions as loaders

Do not break when searching through `use` or `loader` that is a function

* fix(index.js): Support function loaders (#116)

* docs(readme): Tone of Voice clairification on what hot:true does (#113)

* fix(index.js): Support function loaders

updateWebpackConfig previously expected `use` entries to be strings or arrays, and would throw `needle.includes is not a function` errors when given a function that returns a loader object.

* Update README.md

* test: Fixing function-loader tests

* Add reloadAll and cssModules to typings (#115)

* docs(readme): Tone of Voice clairification on what hot:true does (#113)

* Add reloadAll and cssModules to typings